hw1_iris/requirements.txt
numpy==1.24.3
pandas==2.0.0
scikit-learn==1.2.2
joblib==1.2.0
apache-airflow==2.6.1 